Consolidate duplicate utility functions
Extract shared utilities to reduce code duplication:

- port-forward.ts: parsePortForward, formatPortForwards (from proxy.ts, docker-proxy.ts)
- path-utils.ts: expandPath (from config/loader.ts, ssh/sync.ts)

Both modules now re-export for backward compatibility.
