+++ src/client/docker-proxy.ts
@@ -1,9 +1,7 @@
 import type { Socket } from 'bun';
+import { PortForward, parsePortForward, formatPortForwards } from './port-forward';
 
-export interface PortForward {
-  localPort: number;
-  remotePort: number;
-}
+export { PortForward, parsePortForward, formatPortForwards };
 
 export interface DockerProxyOptions {
   containerIp: string;
@@ -100,24 +98,4 @@ export async function startDockerProxy(options: DockerProxyOptions): Promise<() 
       server.stop();
     }
   };
-}
-
-export function formatPortForwards(forwards: PortForward[]): string {
-  return forwards
-    .map((f) =>
-      f.localPort === f.remotePort ? String(f.localPort) : `${f.localPort}:${f.remotePort}`
-    )
-    .join(', ');
-}
-
-export function parsePortForward(spec: string): PortForward {
-  if (spec.includes(':')) {
-    const [local, remote] = spec.split(':');
-    return {
-      localPort: parseInt(local, 10),
-      remotePort: parseInt(remote, 10),
-    };
-  }
-  const port = parseInt(spec, 10);
-  return { localPort: port, remotePort: port };
 }
