+++ src/client/proxy.ts
@@ -1,9 +1,7 @@
 import { spawn } from 'child_process';
+import { PortForward, parsePortForward, formatPortForwards } from './port-forward';
 
-export interface PortForward {
-  localPort: number;
-  remotePort: number;
-}
+export { PortForward, parsePortForward, formatPortForwards };
 
 export interface ProxyOptions {
   worker: string;
@@ -13,18 +11,6 @@ export interface ProxyOptions {
   onConnect?: () => void;
   onDisconnect?: (code: number) => void;
   onError?: (error: Error) => void;
-}
-
-export function parsePortForward(spec: string): PortForward {
-  if (spec.includes(':')) {
-    const [local, remote] = spec.split(':');
-    return {
-      localPort: parseInt(local, 10),
-      remotePort: parseInt(remote, 10),
-    };
-  }
-  const port = parseInt(spec, 10);
-  return { localPort: port, remotePort: port };
 }
 
 export async function startProxy(options: ProxyOptions): Promise<void> {
@@ -125,12 +111,4 @@ export async function startProxy(options: ProxyOptions): Promise<void> {
       }
     });
   });
-}
-
-export function formatPortForwards(forwards: PortForward[]): string {
-  return forwards
-    .map((f) =>
-      f.localPort === f.remotePort ? String(f.localPort) : `${f.localPort}:${f.remotePort}`
-    )
-    .join(', ');
 }
